File replaced by SDT
svn path=/trunk/; revision=8991

+++ libraries/ard/ard1_linmp.f
@@ -73,8 +73,8 @@
 
 *  Arguments Given:
       INTEGER MAP
-      INTEGER DLBND( * )
-      INTEGER DUBND( * )
+      DOUBLE PRECISION DLBND( * )
+      DOUBLE PRECISION DUBND( * )
 
 *  Arguments Returned:
       LOGICAL LINEAR
